embedding_service/requirements.txt
vllm>=0.9.1
fastapi>=0.115.14
uvicorn>=0.35.0
python-dotenv>=1.0.0
torch>=2.1.0
transformers>=4.53.1
numpy>=1.24.0
pydantic>=2.11.7
